refactor: api/common 내부 함수들 api로 위치 변경
