+++ src/hooks/useNotification.tsx
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { createNotification } from '@/api/common/Notification';
+import { createNotification } from '@/api/Notification';
 
 export const useNotification = () => {
   return useMutation({
