+++ src/hooks/useSearch.tsx
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
+import { searchAllResults } from '@/api/Search';
 import type { Post } from '@/type/Post';
 import type { SearchParams } from '@/type/search';
 import type { User } from '@/type/User';
-import { searchAllResults } from '@api/common/Search';
 
 const useSearch = ({ keyword }: SearchParams) => {
   const { data, isFetching, isSuccess } = useQuery<(User | Post)[]>(
