+++ src/pages/ProfilePage.tsx
@@ -3,8 +3,8 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
 import { BiSolidUser } from 'react-icons/bi';
 import { HiPencil } from 'react-icons/hi';
+import { fetchUser } from '@/api/User';
 import { getOptimizedImageURL } from '@/utils/imageURL';
-import { fetchUser } from '@api/common/User';
 import BackButton from '@components/BackButton';
 import BottomNavigation from '@components/BottomNavigation';
 import ScrollToTopButton from '@components/ScrollToTopButton';
